feat(tosu): ✨ 更新数据模型和处理逻辑
将 Score 属性类型从 int 更改为 long，以支持更大的分数值。
删除了不再使用的 TosuDataModels.cs 文件，并更新了 TosuDataSource.cs 中的数据处理逻辑，以适应新的 V2Response 结构。
此外，调整了相关属性的类型以确保一致性。

+++ KeyAsio.MemoryReading/MemoryReadObject.cs
@@ -24,7 +24,7 @@ public partial class MemoryReadObject : INotifyPropertyChanged
     public int Combo { get; set; } // Player.(RulesetPlayData.Combo)
 
     [OnChangedMethod(nameof(OnScoreChanged))]
-    public int Score { get; set; } // Player.(RulesetPlayData.Score)
+    public long Score { get; set; } // Player.(RulesetPlayData.Score)
 
     [OnChangedMethod(nameof(OnIsReplayChanged))]
     public bool IsReplay { get; set; } // Player.IsReplay
